+++ config.py
@@ -1387,17 +1387,21 @@ if bem_mri_images not in ('FLASH', 'T1', 'auto'):
     raise ValueError(msg)
 
 
-def check_baseline(baseline: Tuple[Optional[float], Optional[float]],
-                   epochs_tmin: float, epochs_tmax: float) -> None:
+def check_baseline(
+    *,
+    baseline: Tuple[Optional[float], Optional[float]],
+    epochs_tmin: float,
+    epochs_tmax: float
+) -> None:
     """Raises error if baseline not compatible with [epochs_tmin, epochs_tmax].
 
     Parameters
     ----------
     baseline
         Tuple indicating the beginning and end of the baseline interval.
-    epochs_tmin:
+    epochs_tmin
         Beginning of Epochs.
-    epochs_tmin:
+    epochs_tmax
         End of Epochs.
 
     Raises
@@ -1406,22 +1410,18 @@ def check_baseline(baseline: Tuple[Optional[float], Optional[float]],
         if baseline not contained in [epochs_tmin, epochs_tmax].
         if baseline is not a correct time-interval.
     """
-    if baseline[0] is not None and baseline[0] < epochs_tmin:
-        msg = ('baseline should be contained in [epochs_tmin, epochs_tmax]. '
-               f'But {baseline} in not contained in '
+    if ((baseline[0] is not None and baseline[0] < epochs_tmin) or
+            (baseline[1] is not None and baseline[1] > epochs_tmax)):
+        msg = (f'baseline {baseline} outside of epochs interval '
                f'{[epochs_tmin, epochs_tmax]}.')
-        raise ValueError(msg)
 
-    if baseline[1] is not None and baseline[1] > epochs_tmax:
-        msg = ('baseline should be contained in [epochs_tmin, epochs_tmax]. '
-               f'But {baseline} in not contained in '
-               f'{[epochs_tmin, epochs_tmax]}.')
         raise ValueError(msg)
 
     if ((baseline[0] is not None) and
-        (baseline[1] is not None) and
+            (baseline[1] is not None) and
             (baseline[0] >= baseline[1])):
-        msg = f'{baseline} is not a correct time-interval.'
+        msg = (f'The end of the baseline period must occur after its start, '
+               f'but you set baseline={baseline}'
         raise ValueError(msg)